Improved the stats
Okay, maybe that's subjective but I tried to make the building cargo acceptance better matching their size.

I also modified the start years so the domes roughly line up with the "Spaceship Earth" project (or whatever the biodome project was) and the small habitats now take the mid-CoViD-19 role of total isolation as inspiration.